refactor(templates): move utils to folder
