+++ templates/Dashboard/GridPanel.tsx
@@ -7,7 +7,7 @@ import {
   HvPanel,
   theme,
 } from "@hitachivantara/uikit-react-core";
-import { LoadingContainer } from "templates/LoadingContainer";
+import { LoadingContainer } from "../utils";
 
 /** A `HvGrid` item + styled `HvPanel` container with a loading `Suspense` boundary */
 export const GridPanel = ({
