+++ templates/utils/index.ts
@@ -1,1 +1,2 @@
+export * from "./LoadingContainer";
 export * from "./usePaginationData";
